// frontend/src/security/types.ts
export interface IAvailability {
  /**
   * Some Menu Items will only appear when the logged in Company is from a specific
   * country. For example Australia has some features that are unique and therefore
   * shouldn't appear for the UK. The string must be a two letter ISO country code
   * to avoid type mismatches with the GraphQL API.
   */
  country?: [string];
  /** Different Security roles will have access to different functionality. Restricting
   * the Menu Items by Security roles, means security becomes a configurable issue,
   * rather than a code issue.
   */
  security?: Array<SecurityRole>;
  /** Currently we can restrict menu items to Companies that are of a certain Business
   * Type. Currently only "HEALTH" is a supported option.
   */
  businessType?: Array<string>;
}

/**
 * This is just combination of IAvailability & LicensePlan
 * This will be used on IRoute, IMenuItem and ITab
 */
export interface IContentSecurity {
  /**
   * Conditions to determine whether this Item is available to the User and
   * the Company.
   */
  availability?: Array<IAvailability>;
  /**
   * Conditions to determine whether user needs to upgrade his plan or not
   * Some tabs/menu/routes will be restricted by the License that the company has. For example, a
   * Company may be in the FREE plan, but the Tab is only available on the PRO or
   * ENTERPRISE license. In that instance the Tab should still appear, but a message
   * will be displayed to the User to upgrade their Company plan
   */
  licenseRequired?: LicensePlan[];
}
// ...
export type IRoute = IContentSecurity & {
  /**
   * This name will be used in header bar title
   */
  name: string;
  /**
   * Path will be used to configure route at specific url
   */
  path: string;
  /**
   * Rendering component for particular route, means that if we ever need to completely
   * replace a page within the application. It becomes as simple as just linking to the
   * new React component. This means theoretically, we can generate our entire application
   * by just configuring the Security Config
   */
  component: React.ElementType;
  /**
   * An array of the Tabs that need to appear when the Menu Item is clicked
   */
  tabs?: Array<ITab>;
  /**
   * Any route prop, this will be used to provide any custom route prop in case we have some special case
   */
  [key: string]: any;
};

/**
 * A Menu item will link to multiple tabs. Tabs are a further classification of
 * functionality within the Kudoo application.
 */
export type ITab = IContentSecurity & {
  name: string;
  /**
   * Linking the Tabs to the React Component, means that if we ever need to completely
   * replace a Tab within the application. It becomes as simple as just linking to the
   * new React component. This means theoretically, we can generate our entire application
   * by just configuring the Security Config
   */
  reactComponent: React.ReactNode | React.Component;
  /**
   * Url associated to the tab, this will be used when user click on tab at that time we will
   * redirect that user to particular url. This will be helpful to assign one url per tab so user can
   * easily open any tab with just url in future or user can even bookmark the url
   */
  url: (param?: any) => string;
  /**
   * Sometimes there's too much functionality in a Tab to fit. In that case we have to
   * split the functionality into SubTabs.
   * basically we can keep this key as a "tabs" so we can do unlimited nesting in tabs
   * there is no need to do only 2 level static nesting
   */
  tabs?: Array<ITab>;
  /**
   * Whether rendered tab is secondary tab or tertiary tab
   */
  isTertiaryTab?: boolean;
  /**
   * Any reactComponent specific prop, this will be used to provide any custom component prop in case we have some special case
   */
  [key: string]: any;
};

export enum SecurityRole {
  admin = 'ADMIN',
  user = 'USER',
  root = 'ROOT',
  owner = 'OWNER',
  healthcareProvider = 'HEALTHCARE PROVIDER',
}

export enum LicensePlan {
  PRO = 'PRO',
  ENTERPRISE = 'ENTERPRISE',
  FREE = 'FREE',
}